requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
huggingface-hub==0.24.2
pandas==1.5.3
numpy==1.26.4
openpyxl==3.1.2
python-dotenv==1.0.1
sentence-transformers==3.0.1
streamlit==1.39.0
together==1.2.13
tqdm==4.66.4
transformers==4.43.3
